+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Resolved an issue where models could be set prematurely. (#1)
 - Improved the prompt for text generation.
 - Close the active conversation when it is deleted.
+- Only update document embeddings when a document is updated.
 
 ## [1.1.0](https://github.com/nivthefox/foundryvtt-aide/releases/tag/1.1.0) - 2024-11-09
 ### Added